rename LiveReplayAction class to TrackingLiveReplayAction
