+++ src/main/java/com/faforever/client/replay/TrackingLiveReplayAction.java
@@ -1,5 +1,5 @@
 package com.faforever.client.replay;
 
-public enum LiveReplayAction {
+public enum TrackingLiveReplayAction {
   NOTIFY_ME, RUN_REPLAY
 }
