+++ src/test/java/com/faforever/client/fx/contextmenu/CancelActionNotifyMeMenuItemTest.java
@@ -4,7 +4,7 @@ import com.faforever.client.builders.GameBeanBuilder;
 import com.faforever.client.domain.GameBean;
 import com.faforever.client.i18n.I18n;
 import com.faforever.client.replay.LiveReplayService;
-import com.faforever.client.replay.LiveReplayAction;
+import com.faforever.client.replay.TrackingLiveReplayAction;
 import com.faforever.client.replay.TrackingLiveReplay;
 import com.faforever.client.test.UITest;
 import org.junit.jupiter.api.Test;
@@ -37,7 +37,7 @@ public class CancelActionNotifyMeMenuItemTest extends UITest {
   @Test
   public void testVisibleItem() {
     GameBean game = GameBeanBuilder.create().defaultValues().get();
-    when(liveReplayService.getTrackingLiveReplay()).thenReturn(Optional.of(new TrackingLiveReplay(game.getId(), LiveReplayAction.NOTIFY_ME)));
+    when(liveReplayService.getTrackingLiveReplay()).thenReturn(Optional.of(new TrackingLiveReplay(game.getId(), TrackingLiveReplayAction.NOTIFY_ME)));
 
     instance.setObject(game);
     assertTrue(instance.isVisible());
@@ -59,7 +59,7 @@ public class CancelActionNotifyMeMenuItemTest extends UITest {
   @Test
   public void testInvisibleItemIfNoTrackingOwnReplay() {
     GameBean game = GameBeanBuilder.create().defaultValues().id(1).get();
-    when(liveReplayService.getTrackingLiveReplay()).thenReturn(Optional.of(new TrackingLiveReplay(2, LiveReplayAction.NOTIFY_ME)));
+    when(liveReplayService.getTrackingLiveReplay()).thenReturn(Optional.of(new TrackingLiveReplay(2, TrackingLiveReplayAction.NOTIFY_ME)));
 
     instance.setObject(game);
     assertFalse(instance.isVisible());
