+++ src/test/java/com/faforever/client/fx/contextmenu/CancelActionRunReplayImmediatelyMenuItemTest.java
@@ -4,7 +4,7 @@ import com.faforever.client.builders.GameBeanBuilder;
 import com.faforever.client.domain.GameBean;
 import com.faforever.client.i18n.I18n;
 import com.faforever.client.replay.LiveReplayService;
-import com.faforever.client.replay.LiveReplayAction;
+import com.faforever.client.replay.TrackingLiveReplayAction;
 import com.faforever.client.replay.TrackingLiveReplay;
 import com.faforever.client.test.UITest;
 import org.junit.jupiter.api.Test;
@@ -39,7 +39,7 @@ public class CancelActionRunReplayImmediatelyMenuItemTest extends UITest {
   @Test
   public void testVisibleItem() {
     GameBean game = GameBeanBuilder.create().defaultValues().get();
-    when(liveReplayService.getTrackingLiveReplay()).thenReturn(Optional.of(new TrackingLiveReplay(game.getId(), LiveReplayAction.RUN_REPLAY)));
+    when(liveReplayService.getTrackingLiveReplay()).thenReturn(Optional.of(new TrackingLiveReplay(game.getId(), TrackingLiveReplayAction.RUN_REPLAY)));
 
     instance.setObject(game);
     assertTrue(instance.isVisible());
@@ -61,7 +61,7 @@ public class CancelActionRunReplayImmediatelyMenuItemTest extends UITest {
   @Test
   public void testInvisibleItemIfNoTrackingOwnReplay() {
     GameBean game = GameBeanBuilder.create().defaultValues().id(1).get();
-    when(liveReplayService.getTrackingLiveReplay()).thenReturn(Optional.of(new TrackingLiveReplay(2, LiveReplayAction.RUN_REPLAY)));
+    when(liveReplayService.getTrackingLiveReplay()).thenReturn(Optional.of(new TrackingLiveReplay(2, TrackingLiveReplayAction.RUN_REPLAY)));
 
     instance.setObject(game);
     assertFalse(instance.isVisible());
